Code refactoring to improve memory allocation

+++ header-files/Ambient.h
@@ -7,7 +7,7 @@
 class Ambient {
     public:
         RGBColor color;
-        float ir;
+        double ir;
         int depth = 1, nThreds;
         Ambient(RGBColor c, float i, int d, int nT): color(c), ir(i), depth(d), nThreds(nT) {}
         ~Ambient() {}
